Time preferences manual add working
