+++ frontend/src/components/_shared/Dropdown.jsx
@@ -19,6 +19,7 @@ const Dropdown = ({dropDownDesc, placeholder, required, options, width, onChange
                 width={width}
                 onChange={onChange}
                 defaultValue={defaultValue ? defaultValue : null}
+                mb="5"
             >
                 {options.map((option) => {
                     return <option value={option} label={option}/>
