+++ frontend/src/components/_shared/index.jsx
@@ -10,7 +10,6 @@ import TextField from "./TextField";
 import ToggleButtonGroup from "./ToggleButtonGroup";
 import ToastSettings from "./ToastSettings";
 import InputNumber from "./InputNumber";
-import MultipleDropdown from "./MultipleDropdown";
 
 export {
     AddButton,
@@ -25,5 +24,4 @@ export {
     ToggleButtonGroup,
     ToastSettings,
     InputNumber,
-    MultipleDropdown
 };
